Moved calendar button to components.
